moving teacher signup types to types folder
