+++ frontend/src/APIClients/AuthAPIClient.ts
@@ -3,7 +3,7 @@ import {
   MutationFunctionOptions,
   OperationVariables,
 } from "@apollo/client";
-import { SchoolMetadata } from "../components/auth/TeacherSignup/types";
+import { SchoolMetadata } from "../types/TeacherSignupTypes";
 import AUTHENTICATED_USER_KEY from "../constants/AuthConstants";
 import { AuthenticatedUser } from "../types/AuthTypes";
 import { setLocalStorageObjProperty } from "../utils/LocalStorageUtils";
