+++ frontend/src/components/auth/TeacherSignup/SelectFormInput.tsx
@@ -7,7 +7,10 @@ import {
 import React from "react";
 
 import { UseFormSetValue, UseFormWatch } from "react-hook-form";
-import { TeacherInput, TeacherSignupForm } from "./types";
+import {
+  TeacherInput,
+  TeacherSignupForm,
+} from "../../../types/TeacherSignupTypes";
 
 interface Option extends OptionBase {
   label: string;
